requirements.txt
Babel==1.3
Django==1.7
Flask==0.10.1
Flask-Babel==0.9
Flask-BasicAuth==0.2.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.3
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.2
Flask-WhooshAlchemy==0.56
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.0
SQLAlchemy==0.9.7
Tempita==0.5.3dev
WTForms==2.0.1
Werkzeug==0.9.6
Whoosh==2.6.0
argparse==1.2.1
blinker==1.3
coverage==4.0a1
decorator==3.4.0
distribute==0.6.24
flipflop==1.0
guess-language==0.2
image==1.3.3
itsdangerous==0.24
pbr==0.10.0
python-openid==2.2.5
pytz==2014.7
six==1.8.0
speaklater==1.3
sqlalchemy-migrate==0.9.2
sqlparse==0.1.13
wsgiref==0.1.2
